Revert Thor dependency to 0.15.2 for TorqueBox 2.2.0
Bundler will not install Padrino from github with the updated Thor
dependency since it clashes with TorqueBox during `bundle install`.

Since the TB dependency is in a generated file created by a Maven
build, I've reverted the version in padrino-core for testing with
TorqueBox. Quick tests seem to be working, but I haven't generated
everything.