Padrino::Mounter searches apps in gems now
If Padrino is loaded using rubygems, this patch enables Padrino::Mounter to load apps from gems. To
allow the reloader what it does, this only searches for the app file, but does not load the
application. Applications in gems are expected to be found in `gem-folder/app/app.rb`. Gems named
after `ApplicationConstant.underscore` will be found. Optionally, they can be prefixed with
`padrino-`.