Add Padrino::Module to orchestrate loading
This adds a Padrino::Module module that orchestrates loading of gems.

Padrino::Module implements the most important methods to properly load code from a gem
(#dependency_paths and #root) and adds a method `gem!` to flag a module as being loaded from a gem.

It also modifies the generators so that now, everything works out of the box.