Make Padrino::Rendering use ActiveSupport::SafeBuffer
This change switches all rendering to use SafeBuffer instead of a raw string.

SafeBuffer is a Subclass of String, so concats and instantiations are cheap.

Concatenating SafeBuffer with a normal String escapes the String. Concatenating
SafeBuffer with another SafeBuffer keeps both intact.

This change requires to extensions: a special template implementation for
Erubis and a port of the XssMods used in HAML. At least the latter can be
dropped once HAML makes them available as a standalone.

All strings returned from #render are considered escaped. Strings can be
marked safe for concatenation using String#html_safe, which turns returns the
String as a SafeBuffer.