Fix bug in Padrino::Mounter.app_constant (#1595)
If the module with the same name as the application class has already been defined,
Padrino::Mounter.app_constant had returned the very module object.

So I intend not to search towards ancestors with second parameter 'false' in 'const_defined?'.
This code is only valid in the Ruby 1.9 or later.