ApplicationWrapper Prerequisite Issue
The Reloader and other classes assume that any mounted Padrino app has "prerequisites" (plural!) set, and uses that to figure out what files need to be reloaded. And, in the documentation, it says to ```MyApp.prerequisites << "somefile.rb"``` if you want to configure your application to reload/require new things. 

This doesn't seem to break an actual Padrino app, as all the internal Padrino code (besides this) uses the plural. However, when I mount (and therefore wrap) my Grape::API application, reloading breaks, because there is no list of pluralized prerequisites to work with. Boo!