Override default values if parameter is present
As reported in bug #2113 if we are using default values in the
controllers, and those values are present in the URL, they are not being
correctly replaced when building the URL.

The reason is we are using the default values as overrides and not
otherwise, the given values to override the default ones. This patch
fixes this behaviour and introduces a new test to validate it.