Handle nil database passwords for create_db and drop_db.
Commit ce7ab0a4c2a606dbeef4212db0e74dba67ba81f7 introduced a bug where
if the database configuration didn't specify a password attempts to
create/drop the database would fail. The underlying cause was that prior
to the above the code used `blank?` to which would catch both nil values
and empty values. The new code uses `empty?` which fails if used on a
nil value.

This patch fixes that by checking for nil in addition to an empty string.

This can be reproduced by creating a brand new app, using postgresql or
mysql, and setting the database configuration to this:

  postgres://localhost/nilpassword_development

And then running `rake sq:create` which will result in this:

  $ rake sq:create
  => Creating database 'nilpassword_development'
  rake aborted!
  NoMethodError: undefined method `empty?' for nil:NilClass
  .../gems/padrino-gen-0.13.3.3/lib/padrino-gen/padrino-tasks/sql-helpers.rb:10:in `create_db'
  .../gems/padrino-gen-0.13.3.3/lib/padrino-gen/padrino-tasks/sequel.rb:52:in `block (2 levels) in <top (required)>'
  .../gems/rake-12.0.0/exe/rake:27:in `<top (required)>'
  .../bin/ruby_executable_hooks:15:in `eval'
  .../bin/ruby_executable_hooks:15:in `<main>'
  Tasks: TOP => sq:create
  (See full trace by running task with --trace)